Changed the oplog test API to be able to indicate the expected exception
As the code is not actually failing when some of these test are
executed, it is expected that mvn install fails on this commit!
