+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/applyOps.json
@@ -27,6 +27,7 @@
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         { 
             "op": "c",
