+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/delete_without_id.json
@@ -3,7 +3,8 @@
         "db1": {
             "col1": [
                 {
-                    "id": 1
+                    "_id": 1,
+                    "aField": 1
                 }
             ]
         }
@@ -12,17 +13,19 @@
         "db1": {
             "col1": [
                 {
-                    "id": 1
+                    "_id": 1,
+                    "aField": 1
                 }
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         {
             "op": "d",
             "ns": "db1.col1",
             "o": {
-                "id": 1
+                "aField": 1
             }
         }
     ]
