+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/insert_arr_id.json
@@ -1,6 +1,4 @@
 {
-    "comment": "this test is not executed because the given error is not the expected",
-    "ignore": true,
     "initialState": {
         "db1": {
             "col1": [
@@ -13,6 +11,7 @@
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         {
             "op": "i",
