+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/insert_without_id.json
@@ -11,6 +11,7 @@
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         {
             "ts": {
