+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/update_doc_id.json
@@ -1,6 +1,4 @@
 {
-    "comment": "this test is not executed because the given error is not the expected",
-    "ignore": true,
     "initialState": {
         "db1": {
             "col1": [
@@ -13,12 +11,15 @@
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         {
             "op": "u",
             "ns": "db1.col1",
             "o2": {
-                "_id": [1]
+                "_id": {
+                    "a" : 1
+                }
             },
             "o": {
                 "$set": {
