+++ engine/mongodb/repl/src/test/resources/com/torodb/mongodb/repl/oplogreplier/update_no_upsert_without_id.json
@@ -3,7 +3,8 @@
         "db1": {
             "col1": [
                 {
-                    "id": 1
+                    "_id": 1,
+                    "aField": 1
                 }
             ]
         }
@@ -12,17 +13,19 @@
         "db1": {
             "col1": [
                 {
-                    "id": 1
+                    "_id": 1,
+                    "aField": 1
                 }
             ]
         }
     },
+    "expectedException": "UnexpectedOplogApplierException",
     "oplog": [
         {
             "op": "u",
             "ns": "db1.col1",
             "o2": {
-                "id": 1
+                "aField": 1
             },
             "o": {
                 "$set": {
